CORE-731 Refactor app launch validation function into separate file.
