+++ src/components/apps/launch/AppLaunchWizard.js
@@ -12,6 +12,7 @@ import { injectIntl } from "react-intl";
 import constants from "./constants";
 import ids from "./ids";
 import messages from "./messages";
+import validate from "./validate";
 
 import AnalysisInfoForm from "./AnalysisInfoForm";
 import { ParamGroupForm, ParamsReview } from "./ParamGroups";
@@ -55,52 +56,6 @@ const formatAnalysisName = (intl, name) =>
               "_"
           )
         : "";
-
-const validate = (values) => {
-    const errors = {};
-    const stepErrors = [];
-
-    if (!values.name) {
-        errors.name = getMessage("required");
-        stepErrors[0] = true;
-    }
-    if (!values.output_dir) {
-        errors.output_dir = getMessage("required");
-        stepErrors[0] = true;
-    }
-
-    if (values.groups) {
-        const groupErrors = [];
-        values.groups.forEach((group, index) => {
-            const paramErrors = [];
-
-            if (group.parameters) {
-                group.parameters.forEach((param, paramIndex) => {
-                    if (!param.value && param.required) {
-                        paramErrors[paramIndex] = {
-                            value: getMessage("required"),
-                        };
-                        stepErrors[1] = true;
-                    }
-                });
-
-                if (paramErrors.length > 0) {
-                    groupErrors[index] = { parameters: paramErrors };
-                }
-            }
-        });
-
-        if (groupErrors.length > 0) {
-            errors.groups = groupErrors;
-        }
-    }
-
-    if (stepErrors.length > 0) {
-        errors.steps = stepErrors;
-    }
-
-    return errors;
-};
 
 const initValues = ({
     intl,
